Convert more dialogs to stack frames
Port over the simple entry point dialogs: discovery, device config and init
