+++ lib/default_dialog_queue.js
@@ -11,9 +11,6 @@
 
 const assert = require('assert');
 
-const discoveryDialog = require('./dialogs/discovery');
-const configDialog = require('./dialogs/config');
-const askAnything = require('./dialogs/ask_anything');
 const permissionGrant = require('./dialogs/permission_grant');
 const { showNotification, showError } = require('./dialogs/notifications');
 
@@ -23,6 +20,9 @@ const { CancellationError } = require('./errors');
 const TopLevelStackFrame = require('./stack-frames/top-level');
 const InitStackFrame = require('./stack-frames/init');
 const ProgramStackFrame = require('./stack-frames/program');
+const AsyncQuestionStackFrame = require('./stack-frames/async-question');
+const DeviceDiscoveryStackFrame = require('./stack-frames/discovery');
+const DeviceConfigStackFrame = require('./stack-frames/device-config');
 
 class QueueItem {
     constructor() {
@@ -164,9 +164,8 @@ QueueItem.Question = class Question extends QueueItem {
     }
 
     async _doDispatch(dlg) {
-        dlg.currentAppId = this.appId;
         // note that we propagate all errors here
-        return askAnything(dlg, this.appId, this.icon, this.type, this.question);
+        return dlg.pushStackFrame(new AsyncQuestionStackFrame(dlg, this.appId, this.icon, this.type, this.question), null);
     }
 };
 
@@ -193,9 +192,9 @@ QueueItem.InteractiveConfigure = class InteractiveConfigure extends QueueItem {
     async _doDispatch(dlg) {
         dlg.currentAppId = null;
         if (this.kind !== null)
-            return configDialog(dlg, this.kind);
+            return dlg.pushStackFrame(new DeviceConfigStackFrame(dlg, this.kind), null);
         else
-            return discoveryDialog(dlg);
+            return dlg.pushStackFrame(new DeviceDiscoveryStackFrame(dlg), null);
     }
 };
