+++ lib/stack-frames/base.js
@@ -148,7 +148,7 @@ module.exports = class DialogStackFrame {
       Display debug messages for this stack frame.
     */
     async debug() {
-        await this._dlg.reply("I'm not in the default state");
+        await this._dlg.reply("I'm in a " + this.constructor.name);
         if (this.expecting === null)
             await this._dlg.reply("I'm not expecting anything");
         else
