+++ lib/stack-frames/result.js
@@ -12,8 +12,8 @@
 const { Intent, ValueCategory } = require('../semantic');
 const DialogStackFrame = require('./base');
 
-const askAnything = require('../dialogs/ask_anything');
 const { showNotification, showError } = require('../dialogs/notifications');
+const AsyncQuestionStackFrame = require('./async-question');
 
 const PAGE_SIZE = 5;
 
@@ -123,7 +123,7 @@ module.exports = class ResultStackFrame extends DialogStackFrame {
             let { item: next, resolve, reject } = await this._app.mainOutput.next();
             if (next.isQuestion) {
                 try {
-                    const value = await askAnything(this._dlg, undefined, next.icon, next.type, next.question);
+                    const value = await this._dlg.pushStackFrame(new AsyncQuestionStackFrame(this._dlg, undefined, next.icon, next.type, next.question), null);
                     resolve(value);
                 } catch(e) {
                     reject(e);
